damage up effect

+++ CharredMana/Assets/Scripts/Effects/SerializableEffect.cs
@@ -6,5 +6,5 @@ public class SerializableEffect
 {
     public EffectType Type;
     public int Duration;
-    public float Amplifier;
+    public int Amplifier;
 }
